==> scripts/1A_setup_data_BEFORE.qmd <==
---
title: "Set Up Data"
format: html
editor: visual
---

## Load packages

```{r}
library(tidyverse) 
library(here) # paths start in project folder
library(coralchlo) # chlorophyll calculations
source(here("scripts/helper_functions.R")) # some helper functions to deal with sample IDs
```

## Before Bleaching

See [here](https://andieich.github.io/coral_chlorophyll_symcounts/6_calculations.html) for a description of functions in the `coralchlo` package.

Read Metadata and calculate area

```{r}
metadata_before <- "https://docs.google.com/spreadsheets/d/1goHa8HiUjp9kzqW0u8Ok0BYo3u-ktFIxxbugbTByDEM/edit?gid=0#gid=0"

dat_overview_before <- read_metadata(metadata_before)
```

Calculate chlorophyll concentrations and normalize per area:

```{r}
dat_chl_before <- normalise_chl_per_area(dat_overview_before,
                                    path_to_biotekfolder = "PhD/coral_physiology/charlotte/photometer_files/before_bleaching", 
                                    is_googledrive = T)
```

Normalize counts per surface area:

```{r}
dat_counts_before <- normalise_counts_per_area(dat_overview_before)
```

The sample code had to be changed due to the high mortality. Update to new format:

```{r}
# Chlorophyll
dat_chl_before <- dat_chl_before %>% 
  # Remove 'charlottes' samples
  filter(nchar(sample_id) == 8) %>% 
  mutate(sample_id = paste0(substr(sample_id, 1, 5),
                            0,
                            substr(sample_id, 6, 8)))

# Counts
dat_counts_before <- dat_counts_before %>% 
  # Remove 'charlottes' samples
  filter(nchar(sample_id) == 8) %>% 
  mutate(sample_id = paste0(substr(sample_id, 1, 5),
                            0,
                            substr(sample_id, 6, 8)))
```

Add sample information from sample ID

```{r}
dat_chl_before <- dat_chl_before %>% 
  get_info()

dat_counts_before <- dat_counts_before %>% 
  get_info()
```

Filter *Acropora* shallow

```{r}
dat_chl_before <- dat_chl_before %>% 
  filter(spec == "A. hya", depth == "5 m")

dat_counts_before <- dat_counts_before %>% 
  filter(spec == "A. hya", depth == "5 m")
```

Overview number of samples

```{r}
dat_chl_before %>% 
  # no duplicates (from replicate chl measurements)
  filter(!duplicated(sample_id)) %>% 
  group_by(site, size) %>% 
  summarise(n = n())
```

Save data

```{r}
write.csv(dat_chl_before, 
          here("data/dat_chl_before.csv"), row.names = F)

write.csv(dat_counts_before, 
          here("data/dat_counts_before.csv"), row.names = F)
```
