==> scripts/2_overview_AFTER.qmd <==
---
title: "Overview Plots AFTER bleaching"
format: html
editor: visual
---

## Load libraries

```{r}
library(here)
library(tidyverse)
source(here("scripts/helper_functions.R")) # some helper functions to deal with sample IDs
```

```{r}
dat_chl_after <-read.csv(here("data/dat_chl_after.csv"))

dat_counts_after <- read.csv(here("data/dat_counts_after.csv"))
```

Summarise data

```{r}
dat_chl_afterS <- dat_chl_after %>% 
  group_by(site, size, sample_id) %>% 
  summarise(mean_chl_tot_per_cm2_colony = mean(chl_tot_per_cm2)) %>% 
  group_by(site, size) %>% 
  summarise(mean_chl_tot_per_cm2 = mean(mean_chl_tot_per_cm2_colony),
            n = n(),
            sd = sd(mean_chl_tot_per_cm2_colony),
            se = sd / sqrt(n)
  )
```

Counts:

```{r}
dat_counts_afterS <- dat_counts_after %>% 
   group_by(site, size, sample_id) %>% 
   summarise(mean_count_per_cm2_colony = mean(count_per_cm2)) %>% 
  group_by(site, size) %>%    summarise(mean_count_per_cm2 = mean(mean_count_per_cm2_colony),
           n = n(),
           sd = sd(mean_count_per_cm2_colony),
           se = sd / sqrt(n)
)
```

Graph:chlorophyll

```{r}

col_size <- c("Adult" = "#D16103",
              "Juvenile" = "#293352")

dat_chl_after %>% 
  ggplot(aes(x = site, col = size,))+
  geom_point(aes( y = chl_tot_per_cm2), 
             position = position_jitterdodge(jitter.width = .2, 
                                             dodge.width = .6),
             alpha = .6)+
  geom_errorbar(data = dat_chl_afterS, 
                aes(x = site,
                    ymin = mean_chl_tot_per_cm2 - se,
                    ymax = mean_chl_tot_per_cm2 + se),
                position = position_dodge(width = .6),
                width = .3,
                show.legend = F)+
  geom_point(data = dat_chl_afterS, 
             aes(x = site, 
                 y = mean_chl_tot_per_cm2), 
             size = 4,
             position = position_dodge(width = .6), 
             shape = 21, 
             fill = "white", 
             stroke = 1, 
             show.legend = F)+
  labs(x = "Site", 
       y = expression("Chlorophyll in µg/"~cm^2), 
       title = "Chlorophyll concentration after the bleaching event in May 2024:")+
  scale_colour_manual(values = col_size, name = "Size")+
  theme_minimal()+
  theme(plot.title = element_text(hjust = 0.5))

```

Graph: Counts

```{r}
dat_counts_afterS <- dat_counts_after %>% 
   group_by(site, size, sample_id) %>% 
   summarise(mean_count_per_cm2_colony = mean(count_per_cm2)) %>% 
  group_by(site, size) %>%    summarise(mean_count_per_cm2 = mean(mean_count_per_cm2_colony),
           n = n(),
           sd = sd(mean_count_per_cm2_colony),
           se = sd / sqrt(n)
)
```

Graph: Counts

```{r}

plot_symcount_after <- dat_counts_after %>% 
  ggplot(aes(x = site, col = size,))+
  geom_point(aes( y = count_per_cm2), 
             position = position_jitterdodge(jitter.width = .2, 
                                             dodge.width = .6),
             alpha = .6)+
  geom_errorbar(data = dat_counts_afterS, 
                aes(x = site,
                  ymin = mean_count_per_cm2 - se,
                  ymax = mean_count_per_cm2 + se),
                position = position_dodge(width = .6),
                width = .3,
                show.legend = F)+
  geom_point(data = dat_counts_afterS, 
             aes(x = site, 
                 y = mean_count_per_cm2), 
             size = 4,
             position = position_dodge(width = .6), 
             shape = 21, 
             fill = "white", 
             stroke = 1, 
             show.legend = F)+
  labs(x = "Site", 
       y = expression("Number of Symbiodiniaceae per"~cm^2), 
       title = "Number of Symbiodiniaceae:")+
  scale_colour_manual(values = col_size, name = "Size")+
  theme_minimal()+
  theme(plot.title = element_text(hjust = 0.5))
plot_symcount_after
```

Chlorophyll per symbiont

```{r}
dat_chl_after_persample <- dat_chl_after %>% 
 group_by(sample_id) %>% 
 summarise(chl_tot_per_cm2 = mean(chl_tot_per_cm2))

dat_counts_after_persample <- dat_counts_after %>%      group_by(sample_id) %>% 
  summarise(count_per_cm2 = mean(count_per_cm2))

dat_chl_count_after <- left_join(dat_chl_after_persample,dat_counts_after_persample, by = "sample_id")

dat_chl_count_after <- dat_chl_count_after %>%
  mutate(chl_per_symb = chl_tot_per_cm2 / count_per_cm2 ) %>%
  get_info()

```

```{r}
dat_chl_count_afterS <- dat_chl_count_after %>% 
  group_by(site, size) %>%  
  summarise(mean_chl_per_symb = mean(chl_per_symb),
           n = n(),
           sd = sd(chl_per_symb),
           se = sd / sqrt(n)
)
```

Graph: Chlorophyll per symbiont

```{r}

plot_chl_per_symb<- dat_chl_count_after %>% 
  ggplot(aes(x = site, col = size,))+
  geom_point(aes( y = chl_per_symb), 
             position = position_jitterdodge(jitter.width = .2, 
                                       dodge.width = .6),
             alpha = .6)+
  geom_errorbar(data = dat_chl_count_afterS, 
                aes(x = site,
                  ymin = mean_chl_per_symb - se,
                  ymax = mean_chl_per_symb + se),
                position = position_dodge(width = .6),
                width = .3, show.legend = F)+

  geom_point(data = dat_chl_count_afterS, 
             aes(x = site, 
                 y = mean_chl_per_symb), 
             size = 4,
             position = position_dodge(width = .6), 
             shape = 21, 
             fill = "white", 
             stroke = 1, 
             show.legend = F)+
  labs(x = "Site", 
       y = expression("Chlorophyll per Symbiodiniaceae"), 
       title = "Chlorophyll per Symbiodiniaceae")+
  scale_colour_manual(values = col_size, name = "Size")+
  theme_minimal()+
  theme(plot.title = element_text(hjust = 0.5))
plot_chl_per_symb
```

Graph : Chlorophyll per symbiont (comparison between the two sites)

```{r}
m_dat_chl_count_after <- glm(chl_per_symb ~ site*size, 
                   data = dat_chl_count_after,
                   family = "Gamma")
  
```

```{r}
library(car)
summary(m_dat_chl_count_after)
Anova(m_dat_chl_count_after)

```

```{r}
m_dat_chl_count_after2<- lm(chl_per_symb ~ site*size, 
                   data = dat_chl_count_after)
```

```{r}
library(car)
summary(m_dat_chl_count_after2)
anova(m_dat_chl_count_after2)

```

Copy data to Word/Excel

```{r}
library(broom)
library(htmlTable)

anova(m_dat_chl_count_after2) %>% 
  tidy() %>% 
  janitor::clean_names() %>% #clean names (no spaces etc)
  mutate(p_value = txtPval(p_value, lim.sig = 10^-4)) %>% #round p values, everything below lim.sig is changed to < 0.0001 (or the appropriate number os 0s). Read in ?txtPval
  mutate(sumsq = format(sumsq, digits = 2),
         meansq = format(meansq, digits = 2),
         statistic = format(statistic, digits = 2)) %>% #additionally to the already formatted p-values (pr_f), add all variables you do not want to have rounded (here, df)
  htmlTable()
```

```{r}
#compare models
AIC(m_dat_chl_count_after, m_dat_chl_count_after2) # m_chl_count1 slightly lower

library(performance)
compare_performance(m_dat_chl_count_after, m_dat_chl_count_after2, rank = TRUE)#chose dat_chl_count_after2
```

```{r}
summary(m_dat_chl_count_after2)#Adjusted R-squared:  0.0333
plot(m_dat_chl_count_after2)
```

```{r}
library(emmeans)
library(multcomp)

em_chl_count_after <- emmeans(m_dat_chl_count_after2, pairwise ~ size*site)

dat_em_chl_count_after <- em_chl_count_after %>% 
  cld(Letters = LETTERS) %>% 
  mutate(group = trimws(.group)) %>% 
  dplyr::select(site, size, group)

dat_chl_count_afterS <- left_join(dat_chl_count_afterS,
                            dat_em_chl_count_after,
                            by = c("site", "size"))
  



```

```{r}
emmeans(lm(chl_per_symb ~ size, 
                   data = dat_chl_count_after),
        pairwise ~ size)

```

```{r}

col_site <- c("E2B" = "#009E73",
              "Temae" = "#E69F00")

plot_chl_count_after <- dat_chl_count_after %>% 
  ggplot(aes(x = size, col = site))+
  geom_point(aes(y = chl_per_symb),
             position = position_jitterdodge(jitter.width = .2,
                                       dodge.width = .6),
             alpha = .6)+
  geom_errorbar(data = dat_chl_count_afterS,
                aes(ymin = mean_chl_per_symb - se,
                  ymax = mean_chl_per_symb + se),
                position = position_dodge(width = .6),
                width = .3,
                show.legend = F)+
  
  geom_point(data = dat_chl_count_afterS,
             aes(y = mean_chl_per_symb),
             size = 4,
             position = position_dodge(width = .6),
             shape = 21,
             fill = "white",
             stroke = 1,
             show.legend = F)+

  labs(x = "Size", 
       y = expression("Chlorophyll per Symbiodiniaceae"), 
       title = "Chlorophyll per Symbiodiniaceae:")+
  scale_colour_manual(values = col_site, name = "Site")+
  theme_minimal()+
  theme(plot.title = element_text(hjust = 0.5))

plot_chl_count_after
```

Graph : Concentration Chlorophylle (comparison between the two sites)

```{r}

```

```{r}
dat_chl_after_mod  <-  dat_chl_after  %>% 
  group_by(sample_id, site, size)  %>% 
  summarise(chl_tot_per_cm2 = mean(chl_tot_per_cm2))


m_dat_chl_after <- glm(chl_tot_per_cm2 ~ site*size, 
                   data = dat_chl_after_mod,
                   family = "Gamma")

#null model
m_dat_chl_after0 <- glm(chl_tot_per_cm2 ~ 1, 
                   data = dat_chl_after_mod,
                   family = "Gamma")


```

```{r}
1 - (logLik(m_dat_chl_after)/logLik(m_dat_chl_after0))

```

```{r}
library(car)
summary(m_dat_chl_after)
Anova(m_dat_chl_after)

```

```{r}
m_dat_chl_after2<- lm(chl_tot_per_cm2 ~ site*size, 
                   data = dat_chl_after_mod)
```

```{r}
library(car)
summary(m_dat_chl_after2)
anova(m_dat_chl_after2)

```

```{r}
AIC(m_dat_chl_after, m_dat_chl_after2) 

library(performance)
compare_performance(m_dat_chl_after, m_dat_chl_after2, rank = TRUE)#chose dat_chl_after
```

```{r}
summary(m_dat_chl_after)#Adjusted R-squared:  
plot(m_dat_chl_after)
```

Stat : Nombre de symbiodiniaceae

```{r}
dat_count_after_mod  <-  dat_count_after  %>% 
  group_by(sample_id, site, size)  %>% 
  summarise(count_per_cm2 = mean(count_per_cm2))


m_dat_count_after <- glm(count_per_cm2 ~ site*size, 
                   data = dat_count_after_mod,
                   family = "Gamma")

#null model
m_dat_count_after0 <- glm(count_per_cm2 ~ 1, 
                   data = dat_count_after_mod,
                   family = "Gamma")


```

```{r}
 1 - (logLik(m_dat_count_after)/logLik(m_dat_count_after0))

```

```{r}
library(car)
summary(m_dat_count_after)
Anova(m_dat_count_after)

```

```{r}
m_dat_count_after2<- lm(count_per_cm2 ~ site*size, 
                   data = dat_count_after_mod)
```

```{r}
library(car)
summary(m_dat_count_after2)
anova(m_dat_count_after2)
```
